Move sansio implementation into primary module
